@@ -4,34 +4,33 @@
   "description": "Advent of Code 2019",
   "main": "index.js",
   "engines": {
-    "node": ">=12.4.0"
+    "node": ">=14.15.0"
   },
   "repository": "https://github.com/josepot/AoC-2019",
   "author": "Josep M Sobrepere",
   "license": "MIT",
   "dependencies": {
     "big-integer": "^1.6.48",
-    "id-permutations": "^1.0.3",
-    "immutable": "^3.8.2",
+    "id-permutations": "^1.0.6",
     "md5": "^2.2.1",
-    "priorityqueuejs": "^1.0.0",
-    "ramda": "^0.26.1",
+    "priorityqueuejs": "^2.0.0",
     "rxjs": "^6.5.4"
   },
   "devDependencies": {
     "@types/priorityqueuejs": "^1.0.1",
-    "@typescript-eslint/eslint-plugin": "^2.19.2",
-    "@typescript-eslint/parser": "^2.19.2",
+    "@typescript-eslint/eslint-plugin": "^4.6.1",
+    "@typescript-eslint/parser": "^4.6.1",
     "babel-eslint": "^10.0.3",
     "cross-env": "^7.0.0",
-    "eslint": "^6.8.0",
+    "eslint": "^7.12.1",
     "eslint-config-ts-node": "^0.1.1",
     "eslint-plugin-import": "^2.20.1",
     "husky": "^4.2.2",
     "lint-staged": "^10.0.7",
-    "prettier": "^1.19.1",
+    "prettier": "^2.1.2",
+    "ts-node": "^9.0.0",
     "tsconfig-paths": "^3.9.0",
-    "typescript": "^3.7.5"
+    "typescript": "^4.0.5"
   },
   "prettier": {
     "semi": false,
