@@ -1,4 +1,4 @@
-import graphDistinctSearch from "../utils/ts/graphDistinctSearch"
+import graphDistinctSearch from "utils/graphDistinctSearch"
 
 type Forward = Map<string, Set<string>>
 type Backwards = Map<string, string>
@@ -10,7 +10,7 @@ interface Node {
 const solution1 = (forward: Forward, backwards: Backwards) => {
   const getOrbitsInTree = (root: string, depth = 0): number =>
     [...(forward.get(root) || [])]
-      .map(x => getOrbitsInTree(x, depth + 1))
+      .map((x) => getOrbitsInTree(x, depth + 1))
       .reduce((a, b) => a + b, depth)
 
   let root: string = backwards.keys().next().value
@@ -26,18 +26,18 @@ const solution2 = (forward: Forward, backwards: Backwards) =>
       node.id === "SAN" ||
       [...(forward.get(node.id) || []), backwards.get(node.id) || ""]
         .filter(Boolean)
-        .map(id => ({
+        .map((id) => ({
           id,
           steps: node.steps + 1,
         })),
     (a: Node, b: Node) => b.steps - a.steps,
   ).steps - 2
 
-export default [solution1, solution2].map(fn => (lines: string[]) => {
+export default [solution1, solution2].map((fn) => (lines: string[]) => {
   const forward: Forward = new Map()
   const backwards: Backwards = new Map()
   lines
-    .map(line => line.split(")"))
+    .map((line) => line.split(")"))
     .forEach(([a, b]) => {
       forward.get(a)?.add(b) || forward.set(a, new Set([b]))
       backwards.set(b, a)
