@@ -1,5 +1,5 @@
-import { intCodeProcessor } from "utils/ts/intCodeGenerator"
+import { intCodeProcessor } from "utils/intCodeGenerator"
 
-export default [1, 2].map(input => ([line]: string) =>
+export default [1, 2].map((input) => ([line]: string) =>
   intCodeProcessor<number>(line, () => {}, input),
 )
